+++ src/domain/customer/entity/customer.spec.ts
@@ -1,4 +1,4 @@
-import Address from "./address";
+import Address from "../value-object/address";
 import Customer from "./customer";
 
 describe("Customer unit tests", () => {
