+++ src/domain/customer/entity/customer.ts
@@ -1,4 +1,4 @@
-import Address from "./address";
+import Address from "../value-object/address";
 
 export default class Customer {
