+++ src/domain/customer/event/customer-address-changed-event.ts
@@ -1,6 +1,6 @@
-import EventInterface from "../event.interface";
+import EventInterface from "../../@shared/event/event.interface";
 
-export default class CustomerCreatedEvent implements EventInterface {
+export default class CustomerAddressChangedEvent implements EventInterface {
   dateTimeOcurred: Date;
   eventData: any;
