+++ src/domain/customer/event/customer-created.event.ts
@@ -1,6 +1,6 @@
-import EventInterface from "../event.interface";
+import EventInterface from "../../@shared/event/event.interface";
 
-export default class ProductCreatedEvent implements EventInterface {
+export default class CustomerCreatedEvent implements EventInterface {
   dateTimeOcurred: Date;
   eventData: any;
