+++ src/domain/customer/event/handler/envia-console-log-1.handler.ts
@@ -1,4 +1,4 @@
-import EventHandlerInterface from "../../event-handler.interface";
+import EventHandlerInterface from "../../../@shared/event/event-handler.interface";
 import CustomerCreatedEvent from "../customer-created.event";
 
 export default class EnviaConsoleLog1Handler implements EventHandlerInterface<CustomerCreatedEvent> {
