+++ src/domain/customer/event/handler/envia-console-log.handler.ts
@@ -1,4 +1,4 @@
-import EventHandlerInterface from "../../event-handler.interface";
+import EventHandlerInterface from "../../../@shared/event/event-handler.interface";
 import CustomerAddressChangedEvent from "../customer-address-changed-event";
 
 export default class EnviaConsoleLogHandler implements EventHandlerInterface<CustomerAddressChangedEvent> {
