+++ src/domain/customer/repository/customer-reposity.interface.ts
@@ -1,4 +1,4 @@
 import Customer from "../entity/customer";
-import RepositoryInterface from "./repository-interface";
+import RepositoryInterface from "../../@shared/repository/repository-interface";
 
 export default interface CustomerRepositoryInterface extends RepositoryInterface<Customer> {}
