+++ src/domain/product/event/product-created.event.ts
@@ -1,6 +1,6 @@
-import EventInterface from "../event.interface";
+import EventInterface from "../../@shared/event/event.interface";
 
-export default class CustomerAddressChangedEvent implements EventInterface {
+export default class ProductCreatedEvent implements EventInterface {
   dateTimeOcurred: Date;
   eventData: any;
