+++ src/domain/product/repository/product-repository.interface.ts
@@ -1,5 +1,5 @@
+import RepositoryInterface from "../../@shared/repository/repository-interface";
 import Product from "../entity/product";
-import RepositoryInterface from "./repository-interface";
 
 export default interface ProductRepositoryInterface
   extends RepositoryInterface<Product> {
