+++ src/infrastructure/repository/customer.repository.spec.ts
@@ -1,6 +1,6 @@
 import { Sequelize } from "sequelize-typescript";
-import Customer from "../../domain/entity/customer"; 
-import Address from "../../domain/entity/address";
+import Customer from "../../domain/customer/entity/customer"; 
+import Address from "../../domain/customer/value-object/address";
 import CustomerModel from "../db/sequelize/model/customer.model";
 import CustomerRepository from "./customer.repository";
