+++ src/infrastructure/repository/order.repository.spec.ts
@@ -1,14 +1,14 @@
 import { Sequelize } from "sequelize-typescript";
-import Customer from "../../domain/entity/customer"; 
-import Address from "../../domain/entity/address";
+import Customer from "../../domain/customer/entity/customer"; 
+import Address from "../../domain/customer/value-object/address";
 import CustomerModel from "../db/sequelize/model/customer.model";
 import CustomerRepository from "./customer.repository";
 import OrderItemModel from "../db/sequelize/model/order-item.model";
 import ProductModel from "../db/sequelize/model/product.model";
 import ProductRepository from "./product.repository";
-import Product from "../../domain/entity/product";
-import OrderItem from "../../domain/entity/order_item";
-import Order from "../../domain/entity/order";
+import Product from "../../domain/product/entity/product";
+import OrderItem from "../../domain/checkout/entity/order_item";
+import Order from "../../domain/checkout/entity/order";
 import OrderRepository from "./order.repository";
 import OrderModel from "../db/sequelize/model/order.model";
