+++ src/infrastructure/repository/order.repository.ts
@@ -1,5 +1,5 @@
-import Order from "../../domain/entity/order";
-import OrderItem from "../../domain/entity/order_item";
+import Order from "../../domain/checkout/entity/order";
+import OrderItem from "../../domain/checkout/entity/order_item";
 import OrderItemModel from "../db/sequelize/model/order-item.model";
 import OrderModel from "../db/sequelize/model/order.model";
